The Build process now has the option to load the csv files at the same time
