+++ src/AppBundle/Command/BuildDatabaseCommand.php
@@ -2,25 +2,35 @@
 
 namespace AppBundle\Command;
 
+use AppBundle\Entity\Calendar;
+use AppBundle\Entity\Server;
+use Doctrine\Common\Persistence\ManagerRegistry;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
+use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Output\ConsoleOutputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
-use Symfony\Component\Console\Input\ArrayInput;
 
 class BuildDatabaseCommand extends ContainerAwareCommand
 {
+
+    private $objectManager;
+
+    public function __construct(ManagerRegistry $objectManager)
+    {
+        $this->objectManager = $objectManager;
+        parent::__construct();
+    }
+
     protected function configure()
     {
         $this
             ->setName('app:buildDatabase')
             ->setDescription('Build the Calendar database')
-            ->addArgument('calendar_file', InputArgument::REQUIRED, 'Full path to calendar csv file')
-            ->addArgument('server_file', InputArgument::REQUIRED, 'Full path to server csv file');
-        ;
+            ->addArgument('calendar_file', InputArgument::OPTIONAL, 'Full path to calendar csv file')
+            ->addArgument('server_file', InputArgument::OPTIONAL, 'Full path to server csv file');;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -60,8 +70,65 @@ class BuildDatabaseCommand extends ContainerAwareCommand
         $greetInput = new ArrayInput($arguments);
         $returnCode = $command->run($greetInput, $output);
 
+        $file = $input->getArgument('calendar_file');
 
+        if ($file !== \null) {
 
+            if ($file === \null || !\file_exists($file)) {
+                $io->error("${file} : does not exists");
+            } else {
+
+                $csv = array_map('str_getcsv', file($file));
+                \array_walk($csv, function (&$a) use ($csv) {
+                    $a = \array_combine($csv[0], $a);
+                });
+                array_shift($csv);
+
+                $io->progressStart(count($csv));
+                foreach ($csv as $row) {
+                    $calendar = new Calendar();
+                    $this->objectManager->getManager()->persist($calendar);
+                    $calendar->setType($row['type']);
+                    $calendar->setEventDate(new \DateTime($row['event_date']));
+                    $this->objectManager->getManager()->flush();
+                    $this->objectManager->getManager()->detach($calendar);
+                    $io->progressAdvance();
+                }
+                $io->progressFinish();
+
+                $io->note('Calendar Load Process Completed');
+            }
+
+            $file = $input->getArgument('server_file');
+
+            if ($file !== \null) {
+
+                if ($file === \null || !\file_exists($file)) {
+                    $io->error("${file} : does not exists");
+                } else {
+
+                    $csv = array_map('str_getcsv', file($file));
+                    \array_walk($csv, function (&$a) use ($csv) {
+                        $a = \array_combine($csv[0], $a);
+                    });
+                    array_shift($csv);
+
+                    $io->progressStart(count($csv));
+                    foreach ($csv as $row) {
+                        $server = new Server();
+                        $this->objectManager->getManager()->persist($server);
+                        $server->setName($row['name']);
+                        $server->setIsDisabled((bool)$row['isDisabled']);
+                        $this->objectManager->getManager()->flush();
+                        $this->objectManager->getManager()->detach($server);
+                        $io->progressAdvance();
+                    }
+                    $io->progressFinish();
+
+                    $io->note('Calendar Load Process Completed');
+                }
+            }
+        }
     }
 
 }
